+++ backend/db/migrations/006_ai_infrastructure.sql
@@ -51,7 +51,7 @@ create index if not exists ai_usage_log_batch_id_idx on public.ai_usage_log (bat
 create table if not exists public.companies (
   id uuid primary key default gen_random_uuid(),
   name varchar(255) not null unique,
-  rating int check (rating >= 1 and rating <= 5),
+  rating numeric(2,1) check (rating >= 1.0 and rating <= 5.0),
   reason text,
   notes text,
   created_at timestamptz not null default now(),
